Sample modification.
