+++ errors.jl
@@ -1,0 +1,1 @@
+const E_ABSTRACT="Abstract method. To be overriden by subtype."
